this commit is for adding tests

+++ controllers/partyController/deleteParty.js
@@ -2,6 +2,14 @@ import executeQuery from '../../db/executeQuery';
 
 const partyController={
     async deleteParty(req, res){
+      const { isAdmin } = req.user;    
+
+      if(!isAdmin){
+        return res.status(401).send({
+          status: res.statusCode,
+          error: 'Unauthorized, Only Admin can access this end-point',
+        });
+      }
             const deleteQuery = 'DELETE FROM parties WHERE id=$1 returning *';
             try {
               const { rows } = await executeQuery.query(deleteQuery, [req.params.id]);
